dvb_Reset() hangs DVBlast when using ASI input.
DVBlast hangs when dvb_Reset() in ASI mode. When using ASI as input there is no
need to reset the device and especially resetting a DVB device is wrong. Therefor
abstract the call to dvb_Reset() and replace it with pf_Reset(). Also the udp input
is adapted to provide a function for pf_Reset(). Both asi_Reset() and udp_Reset()
do nothing.