Makefile: Show what is being done not the full commands.
With this patch the DVBlast compilation output is not
cluttered with the full command lines and makes it easier to
spot warnings and errors. If you want to see what commands are
executed run make V=1 or export V=1 and then make.

The new "clean" output will look like this:

make all
  CC      dvblast.c
  CC      util.c
  CC      dvb.c
  CC      udp.c
  CC      asi.c
  CC      demux.c
  CC      output.c
  CC      en50221.c
  CC      comm.c
  CC      mrtg-cnt.c
  LINK    dvblast
  CC      dvblastctl.c
  LINK    dvblastctl

make install
  INSTALL dvblast.1 -> /usr/local/share/man/man1
  INSTALL dvblast dvblastctl dvblast_mmi.sh -> /usr/local/bin

make uninstall
  RM      /usr/local/bin/dvblast
  RM      /usr/local/bin/dvblastctl
  RM      /usr/local/bin/dvblast_mmi.sh
  RM      /usr/local/share/man/man1/dvblast.1

make clean
  CLEAN   dvblast dvblastctl dvblast.o util.o dvb.o udp.o asi.o demux.o output.o en50221.o comm.o mrtg-cnt.o util.o dvblastctl.o

make archive
  ARCHIVE dvblast-2.0.0.tar.bz2
  -rw-r--r-- 1 gf users 108030 Sep  9 12:12 dvblast-2.0.0.tar.bz2

Signed-off-by: Georgi Chorbadzhiyski <gf@unixsol.org>