Add support for DVB-T/S2 and ISDB multistream.
Generic multistream support was added in Linux 3.6 in commit 287cefd096b.

http://git.kernel.org/gitweb.cgi?p=linux/kernel/git/torvalds/linux.git;a=commitdiff;h=287cefd096b

The way it is added it is backwards compatible with older kernels.

This patch implements mis support in DVBlast. Unfortunately I don't
have mis compatible hardware and a way to test it, so I'm posting
this patch as RFC and RFT.

If your frontend reports FE_CAN_MULTISTREAM (you have to use Linux 3.6
and higher) it'll be invaluable to test if multistream support is
working.