Extend support for up to 16 uncommitted ports.
This will allow support for up to 64 satellites.

There are some 8/10 port DiSEqC switches, that act
as an uncommitted switch or a combination of committed
and uncommitted switches. Later is already supported,
and with this change we support former.

If used switch acts as eg. 10-port uncommitted switch,
one can now tune to a satellite on port 6 using dvblast
switch "-k 6" (tune to sixth uncommitted port).

Of course, using more conventional equipment or a switch
that combines committed and uncommitted switches within
itself, a setup would consist of 4 committed ports located
under first two uncommitted ports, and 2 more committed
ports under third uncommitted port.

This kind of setup is already supported, and tuning to
satellite on "port 6" was/is possible using dvblast
switches "-k 2 -S 2" (tune to second uncommitted port
and on second committed port).